--- openbook_auth/views/top_posts/serializers.py
@@ -1,23 +1,0 @@
-from rest_framework import serializers
-
-from openbook_communities.models import Community
-
-
-class GetTopPostCommunityExclusionSerializer(serializers.Serializer):
-    max_id = serializers.IntegerField(
-        required=False,
-    )
-    count = serializers.IntegerField(
-        required=False,
-        max_value=20
-    )
-
-
-class TopPostExclusionCommunitySerializer(serializers.ModelSerializer):
-    class Meta:
-        model = Community
-        fields = (
-            'id',
-            'name',
-            'title',
-        )
